Fixing of no padding cipher.
If noPadding ciphering was done with C_EncryptUpdate or C_DecryptUpdate then the call failed when the updated part was not a multiple of the block size.
This is wrong; these functions must handle any part length.
The requirement of multiple block size does only affect C_EncryptFinal and C_DecryptFinal; when any of these functions is called then the total length must be a multiple of the block size.
This commit fixes these issues.
Also the prediction for the needed output buffer length that is returned when a crypto function is called with NULL_PTR as input should now be more accurate.
To do a thorough test of all this, the "p11test" has been enhanced to use different part lengths to the update function.