Issue #230: Install p11-kit module file
In order for well-behaved applications to discover and use the PKCS#11
provider module, we should install a p11-kit module file for it.

Allow this to be disabled by configuring with --disable-p11-kit, and do
so in the Jenkins builds since they run as an unprivileged user and
can't install to the standard p11-kit module directory.