Allow config path to be specified in C_Initialize() call
The pReserved field of pInitArgs can be used in a token-specific way, to
indicate the config path. This provides an alternative to setting the
SOFTHSM2_CONF environment variable, which may be non-trivial (or indeed
impossible, if you want to do it from a multi-threaded program before
loading the SoftHSM token).

This is exactly how the NSS softokn module does it, and p11-kit supports
the setting of a string in this field by means of the x-init-reserved:
directive in its module file, thus:

module:libsofthsm2.so
x-init-reserved:SOFTHSM2_CONF=/home/dwmw2/git/openconnect/tests/softhsm2.conf

Obviously it's entirely optional; it overrides the existing environment
variable and DEFAULT_SOFTHSM2_CONF if provided, but doesn't have to be
present.