Eliminate SimpleConfigLoader::getConfigPath()
This cleans up the code after the previous commit. There's no functional
change, and we may even want to merge it *into* the previous commit. But
it's separate for now just for ease of review. It's easier to see what
the previous commit is doing, when we're not also moving chunks of code
around...