Allow the use of a RAW ECpoint with ECDH
Oasis pkcs11-curr-v2.40-cs01 Page 48
2.3.8 EC mechanism parameters
says:

"pPublicData 1:
pointer to other party’s EC public key value. A token MUST be able
to accept this value encoded as a raw octet string (as per section
A.5.2 of [ANSI X9.62]). A token MAY, in addition, support accepting
this value as a DER-encoded ECPoint (as per section E.6 of [ANSI
X9.62]) i.e. the same as a CKA_EC_POINT encoding.

Footnote:
The encoding in V2.20 was not specified and resulted in different
implementations choosing different encodings. Applications relying
only on a V2.20 encoding (e.g. the DER variant) other than the one
specified now (raw) may not work with all V2.30 compliant tokens."

This commit will allow either to be accepted.

Only OSSLUtil was changed. Other places in the code may need to be fixed.

 On branch ECDH-compatability
 Changes to be committed:
	modified:   src/lib/crypto/OSSLUtil.cpp