Split 64bit check into compilation and running
The 64bit check actually tests two things: whether the compiler accepts
-m64 and whether the resulting binary can be run. In the case of
cross-compilation, the first can still be determined but the latter has
to be guessed. Guessing the result of the combined check may result in
compile failure because the compiler doesn't actually accept -m64. With
the split check, we can determine that the compiler doesn't accept it
and guess more specifically that the target system will most likely not
be able to run 64 bit binaries.