CMake: Fix shared library dependencies
The autotools build links libsofthsm.so to OpenSSL, but the CMake one
does not do this. This can be observed with `ldd` and GnuTLS 3.6.5 on
Arch Linux:

    $ LD_DEBUG=all p11tool --provider /.../softhsm/libsofthsm2.so
    ...
     25469:     /.../softhsm/libsofthsm2.so: error: symbol lookup error: undefined symbol: DSA_set0_key (fatal)
    ...
    pkcs11_add_provider: PKCS #11 initialization error.

Fix this by adding OpenSSL as dependency for the softhsm_crypto object
library and subsequently propagate this to the softhsm library. Remove
unused static library files while at it.

This requires CMake 3.12, older versions do not allow targets to be
linked with object libraries via target_link_libraries. (If this version
bump is not desired, a workaround needs to be created for
softhsm_objectstore and softhsm_crypto).