CMake: do not install static libsofthsm2 with ENABLE_STATIC=OFF
Previously static and shared libraries were unconditionally built and
installed. This removes the unused ENABLE_SHARED option and makes the
ENABLE_STATIC functional by not installing libsofthsm2.a.
(softhsm2-static still needs to be built for use by softhsm2-util, etc.)

(--disable-static is a standard autotools option, but CMake does not
have this as the project has to be explicit about building static or
shared libraries.)