Avoid creating duplicate mechanisms
I noticed weird behavior when the library is loaded into longer-running
process, which issues several cycles of C_Initialize & C_Finalize,
where I saw exactly the same amount of duplicate items as the amount
of the calls to C_Initialize.

I was not able to replicated it in a testsuite, probably because of
different way how the memory is handled inside of C++ code and when
called from C code as a PKCS#11 library.

This fix is just making sure that the list is empty before filling it
with new mechanisms. Other possiblity would be to move the clean to
C_Finalize(), which is counterpart to the C_Initialize(), containing
the initialization of this list.