Add "objectstore.umask" configuration option for file/directory creation
This defaults to 0077 to preserve the existing behaviour. It can be set to
0027 to allow users in the same group to backup the token files.

It does not attempt to modify the process umask (because this is a library)
so it can only apply in addition to whatever umask the process has.

Fixes #560 after #40 restricted all new files to the current user.