softhsm2-util: add --show-config parameter
Enables access to build configuration.

Currently implemented the `default-pkcs11-lib` variable to print the default
PKCS#11 library path.

Useful for applications which are using the softhsm2-util and require to
find the provider in filesystem which is used by the utility as there is no
standard process to locate the library.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>