Limit number of sessions
When SoftHSMv2 is used for testing PKCS#11 client code
before hardware is available, maximum session count on
a token needs to be set in order to detects bugs in client
code where sessions are leaked and remain open. This
can cause problems since most hardware implementations
have a limit on opened sessions count.

This change sets a limit to general session count.